Add files via upload
This is an example of Data Augmentation. 
A convolutional neural network that can robustly classify objects even if its placed in different orientations is said to have the property called invariance.
In the real-world scenario, we may have a dataset of images taken in a limited set of conditions. But, our target application may exist in a variety of conditions, such as different orientations, locations, scales, brightness, etc. We account for these situations by training our neural network with additional synthetically modified data.

I have taken a cat image and then applied data augmentation on it. And these are the following outputs.